+++ eslint.config.ts
@@ -1,6 +1,6 @@
 // @ts-check
 import { sxzz } from '@sxzz/eslint-config'
-import vueMacros from './packages/eslint-config/src/flat.ts'
+import vueMacros from './packages/eslint-config/src/index.ts'
 
 export default sxzz([
   vueMacros,