+++ packages/hoist-static/src/api.ts
@@ -1,0 +1,1 @@
+export { transformHoistStatic } from './core'