+++ playground/vue3/vite.js
@@ -19,6 +19,7 @@ startVite(undefined, {
   },
   define: {
     'import.meta.DEV': 'true',
+    TSUP_FORMAT: "'esm'",
   },
   plugins: [Macros(), Raw()],
 })